fix(AI chat): address review feedback
